// src/entities/Credit/controllers/creditController.ts
import type { ICreditController } from "../domain/interfaces/ICreditController.js";
import { CreditRepository } from "../repository/creditRepository.js";
import type { Request, Response } from "express";
import { CreditUseCasesController} from "./creditUseCasesController.js";

export class CreditController implements ICreditController {
  private static instance: CreditController;

  public static getInstance(): CreditController {
    if (!CreditController.instance) {
      CreditController.instance = new CreditController();
    }
    return CreditController.instance;
  }

  private readonly useCases = new CreditUseCasesController(
    CreditRepository.getInstance()
  );

  get = async (req: Request, res: Response): Promise<void> => {
    try {
      const result = await this.useCases.get.execute(Number(req.params.id));
      res.status(200).json(result);
    } catch (error: any) {
      res.status(404).json({ error: error.message });
    }
  };

  getAll = async (req: Request, res: Response): Promise<void> => {
    try {
      const result = await this.useCases.getAll.execute();
      res.status(200).json(result);
    } catch (error: any) {
      res.status(404).json({ error: error.message });
    }
  };

  post = async (req: Request, res: Response): Promise<void> => {
    try {
      const result = await this.useCases.post.execute(req.body);
      res.status(200).json(result);
    } catch (error: any) {
      res.status(403).json({ error: error.message });
    }
  };

  patch = async (req: Request, res: Response): Promise<void> => {
    try {
      const result = await this.useCases.patch.execute(
        Number(req.params.id),
        req.body
      );
      res.status(200).json(result);
    } catch (error: any) {
      res.status(404).json({ error: error.message });
    }
  };

  delete = async (req: Request, res: Response): Promise<void> => {
    try {
      const result = await this.useCases.delete.execute(Number(req.params.id));
      res.status(200).json(result);
    } catch (error: any) {
      res.status(404).json({ error: error.message });
    }
  };
}
